rename packages added dependency
