+++ src/main/java/ru/haazad/onlinestorage/webapp/repositories/CartRepository.java
@@ -1,7 +1,7 @@
-package ru.haazad.onlinestorage.webapp.repository;
+package ru.haazad.onlinestorage.webapp.repositories;
 
 import org.springframework.stereotype.Repository;
-import ru.haazad.onlinestorage.webapp.model.Product;
+import ru.haazad.onlinestorage.webapp.models.Product;
 
 import java.util.List;
