+++ src/main/java/ru/haazad/onlinestorage/webapp/services/ProductService.java
@@ -1,8 +1,8 @@
-package ru.haazad.onlinestorage.webapp.service;
+package ru.haazad.onlinestorage.webapp.services;
 
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
-import ru.haazad.onlinestorage.webapp.model.Product;
+import ru.haazad.onlinestorage.webapp.models.Product;
 
 import java.util.List;
 import java.util.Optional;
